fix: prevent multiple concurrent refreshes for a team
If chunk scanning takes too long, it was possible for a new refresh
to start while another is still in progress, causing excessive disk/cpu
activity and a bunch of warnings in the server log.

Add an in-progress per-team "lock" to prevent this and also increase
the default refresh check interval from 1 minute to 5 minutes.

https://github.com/FTBTeam/FTB-Modpack-Issues/issues/7179